mt6592-common: Add a shim to silence prebuilt audio HAL
For whatever reason, our prebuilt HAL likes to spam the hell out of
the logcat on devices running userdebug builds (which is our case).

It would be cool if this was something more modern (speaking of the
Android version) since shimming the library targeting the original
symbol names would work. Unfortunately, this is not our case and we
need to stub the original symbol names and shim the stubbed symbol
names. We replaced __android_log_print with __android_stfuprint and
__xlog_buf_printf with __xlog_stfuprintf [1].

[1] https://github.com/bq-mt65x2-dev/vendor_bq/commit/7a99bf1bebd1de30e7b16f33aa42878d05fbbca2

Co-authored-by: Jan Engelmohr <jan.engelmohr@mailbox.tu-dresden.de>